Remove need for goawk_context build tag
It's only 1% slower, and the other recent improvements significantly
exceed this, so the next release will be faster in any case. I think
it's worth it to avoid a build tag.
